requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0

# LLM and API
openai>=1.0.0

# Dataset loading
datasets>=2.0.0

# Text processing
num2words>=0.5.12
pyyaml>=6.0
unidecode>=1.3.0
jiwer>=3.0.0

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.0.0
